Memory optimization: float32 construction + numpy panel + window cap
Peak memory drops from ~21+ GB to ~14-16 GB without reducing features.

float32 at creation (model.py + alt_features.py):
- Added _f32() helper, wrapped all ~50+ feature assignments
- Alt-features _merge() now casts to float32 at reindex time
- Impact: every feature DataFrame is 66 MB (was 132 MB)

numpy pre-allocation replaces .stack() + pd.concat (model.py):
- Old: 206 x .stack() → 206 Series (57 GB) → pd.concat (28 GB copy)
- New: pre-allocate np.empty float32 array, features.pop() frees each
  DataFrame as it's consumed, pd.MultiIndex.from_product for index
- Row order verified: ravel(C) matches from_product([dates, tickers])
- Impact: eliminates 57 GB intermediate + 28 GB concat copy

max_train_days=756 moved into fit_predict (model.py):
- Was only applied inside _train_ensemble (too late — boolean slice
  already created the full expanding-window copy)
- Now truncates train_dates BEFORE boolean indexing
- Bounds largest window: 756d × 5K tickers × 206 × 4B = 3.2 GB
  (was up to 14 GB for last expanding window)
- inf-replacement done ONCE at top of fit_predict, not per-window

Free raw alt-data after build_alt_features (run_strategy.py):
- del edgar_df, fred_df, vix_df + 12 other raw source DataFrames
- gc.collect() frees ~2 GB before feature construction starts

+++ src/alt_features.py
@@ -1451,7 +1451,7 @@ def build_alt_features(
     def _merge(d: dict):
         for k, v in d.items():
             if v is not None and not v.empty:
-                all_signals[k] = v.reindex(index=date_index, columns=tickers, method="ffill")
+                all_signals[k] = v.reindex(index=date_index, columns=tickers, method="ffill").astype(np.float32)
 
     _merge(build_fundamental_signals(
         edgar_df, tickers, date_index, sector_map=sector_map, close=close,
